Centralize building-block loading and fix Windows permission test
