+++ synplan/utils/loading.py
@@ -1,12 +1,8 @@
 """Module containing functions for loading reaction rules, building blocks and
 retrosynthetic models."""
 
-import contextlib
-import csv
 import functools
-import gzip
 import logging
-import os
 import pickle
 import shutil
 import warnings
@@ -18,33 +14,19 @@ import yaml
 from chython import smarts as smarts_parser
 from chython.containers import ReactionContainer
 from chython.files.daylight.tokenize import smarts_tokenize
-from chython.files.SDFrw import SDFRead
 from chython.reactor.reactor import Reactor
 from huggingface_hub import hf_hub_download, snapshot_download
-from tqdm.auto import tqdm
 
+from synplan.chem.building_blocks.io import load_building_blocks as load_building_blocks
 from synplan.chem.reaction import CanonicalRetroReactor
 from synplan.chem.reaction.config import ReactorConfig
 from synplan.chem.reaction.rules.symmetry import needs_decollapsed_matches
 from synplan.chem.utils import (
     AtomMappingCheck,
     reaction_string_mapping_status,
-    standardize_sdf_text,
-    standardize_smiles_batch,
 )
-from synplan.utils.files import (
-    count_sdf_records,
-    count_smiles_records,
-    iter_csv_smiles,
-    iter_csv_smiles_blocks,
-    iter_sdf_text_blocks,
-    iter_smiles,
-    iter_smiles_blocks,
-)
-from synplan.utils.parallel import process_pool_map_stream
 
 if TYPE_CHECKING:
-    from synplan.chem.building_blocks import BuildingBlockCatalogue
     from synplan.mcts.config import CombinedPolicyConfig
     from synplan.mcts.evaluation import (
         EvaluationStrategy,
@@ -61,36 +43,6 @@ REPO_ID = "Laboratoire-De-Chemoinformatique/SynPlanner-data"
 LEGACY_REPO_ID = "Laboratoire-De-Chemoinformatique/SynPlanner"
 _MAX_REGULAR_ATOM_NUMBER = 10**9
 logger = logging.getLogger(__name__)
-
-
-def _building_blocks_progress(total: int | None, *, silent: bool):
-    """Create a consistent progress bar for building blocks loading."""
-    if silent:
-        return None
-    return tqdm(
-        total=total,
-        desc="Building blocks",
-        unit="mol",
-        unit_scale=True,
-        unit_divisor=1000,
-        dynamic_ncols=True,
-        smoothing=0.1,
-        disable=silent,
-    )
-
-
-def _map_blocks(blocks, worker_fn, *, num_workers: int):
-    """Map blocks through worker function, optionally using a process pool.
-
-    For `num_workers == 1`, this runs sequentially to avoid process-spawn overhead.
-    """
-    if num_workers < 1:
-        raise ValueError("num_workers must be >= 1")
-    if num_workers == 1:
-        for block in blocks:
-            yield worker_fn(block)
-        return
-    yield from process_pool_map_stream(blocks, worker_fn, max_workers=num_workers)
 
 
 def _extract_zip(zip_path: Path, out_dir: Path) -> None:
@@ -481,164 +433,6 @@ def _load_rules_pickle(file: str) -> tuple[CanonicalRetroReactor, ...]:
         reaction_rules = [rule for rule, _ in reaction_rules]
 
     return tuple(reaction_rules)
-
-
-def load_building_blocks(
-    building_blocks_path: str | Path,
-    standardize: bool = True,
-    silent: bool = True,
-    num_workers: int | None = None,
-    chunksize: int = 1000,
-    *,
-    header: bool = True,
-    delimiter: str = ",",
-    smiles_column: str = "SMILES",
-) -> "frozenset[str] | BuildingBlockCatalogue":
-    """Load molecular stock or a cached stereo-aware vendor catalogue.
-
-    Prepared catalogues retain their InChIKeys and vendor offers and skip
-    chemistry preparation, regardless of ``standardize``.
-    JSON/JSON.GZ is indexed into SQLite once; vendor TSV/TSV.GZ first performs
-    chemistry preparation. SQLite files open directly with bounded reader caches.
-
-    :param building_blocks_path: The path to the file containing the building blocks.
-    :param standardize: Flag if building blocks have to be standardized before loading. Default=True.
-    :param header: For CSV/CSV.GZ files: treat the first row as header. Default=True.
-    :param delimiter: For CSV/CSV.GZ files: delimiter character. Default=",".
-    :param smiles_column: For CSV/CSV.GZ files: header column name containing SMILES.
-        Default="SMILES" (case-insensitive match is supported).
-    :return: A read-only catalogue for JSON/SQLite/vendor TSV, otherwise a SMILES set.
-    """
-
-    building_blocks_path = Path(building_blocks_path).resolve()
-    suffixes = "".join(building_blocks_path.suffixes).lower()
-    vendor_tsv = False
-    if header and suffixes.endswith((".tsv", ".tsv.gz")):
-        opener = gzip.open if suffixes.endswith(".gz") else open
-        with opener(building_blocks_path, "rt", encoding="utf-8", newline="") as stream:
-            columns = next(csv.reader(stream, delimiter="\t"), ())
-        vendor_tsv = any(column.casefold().endswith("_ppg") for column in columns)
-    if suffixes.endswith((".json", ".json.gz", ".sqlite")) or vendor_tsv:
-        from synplan.chem.building_blocks import load_building_block_catalogue
-
-        return load_building_block_catalogue(
-            building_blocks_path, num_workers=1 if num_workers is None else num_workers
-        )
-    return _load_smiles_building_blocks(
-        building_blocks_path,
-        standardize,
-        silent,
-        num_workers,
-        chunksize,
-        header=header,
-        delimiter=delimiter,
-        smiles_column=smiles_column,
-    )
-
-
-@functools.cache
-def _load_smiles_building_blocks(
-    building_blocks_path,
-    standardize,
-    silent,
-    num_workers,
-    chunksize,
-    *,
-    header,
-    delimiter,
-    smiles_column,
-):
-    suffixes = "".join(building_blocks_path.suffixes).lower()
-    is_csv = suffixes.endswith(".csv") or suffixes.endswith(".csv.gz")
-    is_tsv = suffixes.endswith(".tsv") or suffixes.endswith(".tsv.gz")
-    if is_tsv:
-        is_csv = True
-        delimiter = "\t"
-    suffix = building_blocks_path.suffix.lower()
-    if not is_csv and suffix not in {".smi", ".smiles", ".sdf"}:
-        raise ValueError(
-            f"Unsupported building blocks file extension: '{building_blocks_path.name}'. "
-            "Supported: .smi, .smiles, .sdf, .csv, .csv.gz, .tsv, .tsv.gz, .json, .json.gz"
-        )
-
-    building_blocks_smiles = set()
-    if standardize:
-        if num_workers is None:
-            num_workers = max(1, os.cpu_count() - 1)
-        if num_workers < 1:
-            raise ValueError("num_workers must be >= 1")
-
-        if suffix in {".smi", ".smiles"}:
-            total = count_smiles_records(building_blocks_path) if not silent else None
-            step = max(1, chunksize or 1000)
-
-            progress_iter = _building_blocks_progress(total, silent=silent)
-            for out in _map_blocks(
-                iter_smiles_blocks(building_blocks_path, step),
-                standardize_smiles_batch,
-                num_workers=num_workers,
-            ):
-                if out:
-                    building_blocks_smiles.update(out)
-                    if progress_iter is not None:
-                        progress_iter.update(len(out))
-            if progress_iter is not None:
-                progress_iter.close()
-
-        elif is_csv:
-            step = max(1, chunksize or 1000)
-            progress_iter = _building_blocks_progress(None, silent=silent)
-            blocks = iter_csv_smiles_blocks(
-                building_blocks_path,
-                step,
-                header=header,
-                delimiter=delimiter,
-                smiles_column=smiles_column,
-            )
-            for out in _map_blocks(
-                blocks, standardize_smiles_batch, num_workers=num_workers
-            ):
-                if out:
-                    building_blocks_smiles.update(out)
-                    if progress_iter is not None:
-                        progress_iter.update(len(out))
-            if progress_iter is not None:
-                progress_iter.close()
-
-        elif suffix == ".sdf":
-            n = count_sdf_records(building_blocks_path) if not silent else None
-            step = max(1, chunksize or 5000)
-            blocks = iter_sdf_text_blocks(building_blocks_path, step)
-
-            progress = _building_blocks_progress(n, silent=silent)
-            for chunk_out in _map_blocks(
-                blocks, standardize_sdf_text, num_workers=num_workers
-            ):
-                if chunk_out:
-                    building_blocks_smiles.update(chunk_out)
-                    if progress is not None:
-                        progress.update(len(chunk_out))
-            if progress is not None:
-                progress.close()
-    else:
-        if suffix in {".smi", ".smiles"}:
-            for smiles in iter_smiles(building_blocks_path):
-                building_blocks_smiles.add(smiles)
-        elif is_csv:
-            for smiles in iter_csv_smiles(
-                building_blocks_path,
-                header=header,
-                delimiter=delimiter,
-                smiles_column=smiles_column,
-            ):
-                building_blocks_smiles.add(smiles)
-        elif suffix == ".sdf":
-            with SDFRead(str(building_blocks_path)) as sdf:
-                for mol in sdf:
-                    with contextlib.suppress(Exception):
-                        building_blocks_smiles.add(str(mol))
-
-    return frozenset(building_blocks_smiles)
 
 
 def load_value_net(
