+++ tests/unit/utils/test_loading.py
@@ -1,17 +1,26 @@
 import gzip
 
+import pytest
+
+from synplan.chem.building_blocks import load_building_blocks as catalogue_loader
 from synplan.chem.utils import standardize_smiles_batch
 from synplan.utils.loading import load_building_blocks, load_policy_function
 
 
-def test_load_building_blocks_csv_header(tmp_path):
-    path = tmp_path / "bbs.csv"
-    path.write_text(
-        "SMILES,ID\nCCO,1\n,2\nCCO,3\nCCN,4\n",
-        encoding="utf-8",
+@pytest.mark.parametrize("extension", [".csv", ".csv.gz", ".tsv", ".tsv.gz"])
+@pytest.mark.parametrize("standardize", [False, True])
+def test_load_building_blocks_table_header(tmp_path, extension, standardize):
+    assert load_building_blocks is catalogue_loader
+    path = tmp_path / ("bbs" + extension)
+    data = "SMILES,ID\nCCO,1\n,2\nCCO,3\nCCN,4\n"
+    if extension.startswith(".tsv"):
+        data = data.replace(",", "\t")
+    path.write_bytes(
+        gzip.compress(data.encode()) if extension.endswith(".gz") else data.encode()
     )
 
-    bbs = load_building_blocks(path, standardize=False, silent=True)
+    bbs = load_building_blocks(path, standardize=standardize, num_workers=1)
+    assert isinstance(bbs, frozenset)
     assert bbs == frozenset({"CCO", "CCN"})
 
 
@@ -24,9 +33,11 @@ def test_load_building_blocks_csv_header_case_insensitive_column(tmp_path):
     assert bbs == frozenset({"CCO", "CCN"})
 
 
-def test_load_building_blocks_csv_no_header(tmp_path):
-    path = tmp_path / "bbs.csv"
-    path.write_text("CCO,1\nCCN,2\n\n", encoding="utf-8")
+@pytest.mark.parametrize("extension", [".csv", ".tsv"])
+def test_load_building_blocks_table_no_header(tmp_path, extension):
+    path = tmp_path / ("bbs" + extension)
+    delimiter = "\t" if extension == ".tsv" else ","
+    path.write_text(f"CCO{delimiter}1\nCCN{delimiter}2\n\n", encoding="utf-8")
 
     bbs = load_building_blocks(path, standardize=False, silent=True, header=False)
     assert bbs == frozenset({"CCO", "CCN"})
@@ -41,12 +52,14 @@ def test_load_building_blocks_csv_gz(tmp_path):
     assert bbs == frozenset({"CCO", "CCN"})
 
 
-def test_load_building_blocks_csv_standardize_true_runs(tmp_path):
-    path = tmp_path / "bbs.csv"
+@pytest.mark.parametrize("extension", [".csv", ".tsv"])
+@pytest.mark.parametrize("workers", [1, 2])
+def test_load_building_blocks_table_standardization(tmp_path, extension, workers):
+    path = tmp_path / ("bbs" + extension)
     path.write_text("SMILES\nOCC\nCCN\n", encoding="utf-8")
 
     expected = frozenset(standardize_smiles_batch(["OCC", "CCN"]))
-    bbs = load_building_blocks(path, standardize=True, silent=True, num_workers=1)
+    bbs = load_building_blocks(path, standardize=True, silent=True, num_workers=workers)
     assert bbs == expected
